Refactor : 어드민 도메인 수정
